+++ packages/nimble-components/src/anchor-tree-item/tests/anchor-tree-item.spec.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-classes-per-file */
 import { customElement, html, ref } from '@microsoft/fast-element';
 import { TreeItem as FoundationTreeItem } from '@microsoft/fast-foundation';
 import { parameterizeSpec } from '@ni/jasmine-parameterized';