+++ packages/nimble-tokens/.eslintrc.js
@@ -1,25 +1,23 @@
 module.exports = {
     root: true,
-    extends: [
-        '@ni/eslint-config-javascript'
+    ignorePatterns: [
+        'node_modules',
+        'dist',
+        'dist/styledictionary/*',
     ],
     overrides: [
         {
-            parserOptions: {
-                ecmaVersion: '2020'
-            },
-            files: [
-                'source/styledictionary/*.js',
-            ],
+            files: ['*.js'],
+            extends: ['@ni-private/eslint-config-nimble/javascript'],
+        },
+        {
+            files: ['source/styledictionary/*.js'],
             rules: {
                 // Build scripts will not be in published package and are allowed to use devDependencies
                 'import/no-extraneous-dependencies': ['error', { devDependencies: true }],
 
                 // Okay to use console.log in build scripts
                 'no-console': 'off',
-
-                // Enabled to prevent accidental usage of async-await
-                'require-await': 'error'
             }
         }
     ]