+++ packages/performance/.eslintrc.cjs
@@ -1,37 +1,28 @@
 module.exports = {
     root: true,
-    extends: [
-        '@ni/eslint-config-javascript'
-    ],
     ignorePatterns: [
-        // Force inclusion of config dot file
-        '!.eslintrc.js',
         'node_modules',
         'dist'
     ],
-    rules: {
-        // Enabled to prevent accidental usage of async-await
-        'require-await': 'error'
-    },
-    overrides: [{
-        files: ['*.ts'],
-        extends: [
-            '@ni/eslint-config-typescript',
-            '@ni/eslint-config-typescript/requiring-type-checking'
-        ],
-        parserOptions: {
-            project: './tsconfig.json',
-            tsconfigRootDir: __dirname
+    overrides: [
+        {
+            files: ['*.js', '*.cjs'],
+            extends: ['@ni-private/eslint-config-nimble/javascript'],
         },
-        rules: {
-            // Rules enabled due to strictNullChecks
-            '@typescript-eslint/no-non-null-assertion': 'off',
+        {
+            files: ['*.ts'],
+            extends: ['@ni-private/eslint-config-nimble/typescript'],
+            parserOptions: {
+                project: './tsconfig.json',
+                tsconfigRootDir: __dirname
+            }
+        },
+        {
+            files: ['vite.config.js'],
+            rules: {
+                // Configuration scripts will not be in published package and are allowed to use devDependencies
+                'import/no-extraneous-dependencies': ['error', { devDependencies: true }],
+            }
         }
-    }, {
-        files: ['vite.config.js'],
-        rules: {
-            // Configuration scripts will not be in published package and are allowed to use devDependencies
-            'import/no-extraneous-dependencies': ['error', { devDependencies: true }],
-        }
-    }]
+    ]
 };