+++ packages/storybook/.storybook/blocks/.eslintrc.js
@@ -1,0 +1,11 @@
+module.exports = {
+    overrides: [
+        {
+            files: ['*.tsx'],
+            parserOptions: {
+                project: './tsconfig.json',
+                tsconfigRootDir: __dirname
+            },
+        }
+    ]
+};